vitalscan/backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
redis
python-dotenv
pydantic
